create state-managing user location listener with coordinate logging
